+++ src/aks-preview/azext_aks_preview/agentpool_decorator.py
@@ -384,6 +384,25 @@ class AKSPreviewAgentPoolContext(AKSAgentPoolContext):
             ):
                 enable_artifact_streaming = self.agentpool.artifact_streaming_profile.enabled
         return enable_artifact_streaming
+    
+    def get_pod_ip_allocation_mode(self, read_only: bool = False) -> Union[str, None]:
+        """Get the value of pod_ip_allocation_mode.
+        :return: str or None
+        """
+        # overwrite if provided by user
+        pod_ip_allocation_mode = self.raw_param.get("pod_ip_allocation_mode")
+        # In create mode, try to read the property value corresponding to the parameter from the `agentpool` object
+        # if the agentpool property is specified then the property value will be used and cannot be mutated
+        if self.decorator_mode == DecoratorMode.CREATE:
+            if (
+                self.agentpool and
+                self.agentpool.network_profile is not None and
+                self.agentpool.network_profile.pod_ip_allocation_mode is not None
+            ):
+                pod_ip_allocation_mode = self.agentpool.network_profile.pod_ip_allocation_mode
+
+        return pod_ip_allocation_mode
+    
 
     def get_ssh_access(self) -> Union[str, None]:
         """Obtain the value of ssh_access.